Fix CFX_AndroidFontInfo::GetFontData return type.
Make the return type match that of its parent interface,
which is now a size_t.

PiperOrigin-RevId: 432285132
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/91191
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ core/fxge/android/cfx_androidfontinfo.cpp
@@ -55,9 +55,9 @@ void* CFX_AndroidFontInfo::GetFont(const ByteString& face) {
   return nullptr;
 }
 
-uint32_t CFX_AndroidFontInfo::GetFontData(void* hFont,
-                                          uint32_t table,
-                                          pdfium::span<uint8_t> buffer) {
+size_t CFX_AndroidFontInfo::GetFontData(void* hFont,
+                                        uint32_t table,
+                                        pdfium::span<uint8_t> buffer) {
   if (!hFont)
     return 0;
   return static_cast<CFPF_SkiaFont*>(hFont)->GetFontData(table, buffer);
